+++ bin/node/src/commands/block_producer.rs
@@ -1,10 +1,8 @@
-use std::sync::Arc;
 use std::time::Duration;
 
 use anyhow::Context;
 use miden_node_block_producer::BlockProducer;
 use miden_node_utils::grpc::UrlExt;
-use tokio::sync::Barrier;
 use url::Url;
 
 use super::{ENV_BLOCK_PRODUCER_URL, ENV_STORE_BLOCK_PRODUCER_URL};
@@ -93,7 +91,6 @@ impl BlockProducerCommand {
             block_interval: block_producer.block_interval,
             max_txs_per_batch: block_producer.max_txs_per_batch,
             max_batches_per_block: block_producer.max_batches_per_block,
-            production_checkpoint: Arc::new(Barrier::new(1)),
             grpc_timeout,
             mempool_tx_capacity: block_producer.mempool_tx_capacity,
         }