+++ crates/block-producer/src/server/mod.rs
@@ -18,7 +18,7 @@ use miden_protocol::block::BlockNumber;
 use miden_protocol::transaction::ProvenTransaction;
 use miden_protocol::utils::serde::Deserializable;
 use tokio::net::TcpListener;
-use tokio::sync::{Barrier, Mutex, RwLock};
+use tokio::sync::{Mutex, RwLock};
 use tokio_stream::wrappers::{ReceiverStream, TcpListenerStream};
 use tonic::Status;
 use tower_http::trace::TraceLayer;
@@ -65,11 +65,6 @@ pub struct BlockProducer {
     pub max_txs_per_batch: usize,
     /// The maximum number of batches per block.
     pub max_batches_per_block: usize,
-    /// Block production only begins after this checkpoint barrier completes.
-    ///
-    /// The block-producers gRPC endpoint will be available before this point, so this lets the
-    /// mempool synchronize its event stream without risking a race condition.
-    pub production_checkpoint: Arc<Barrier>,
     /// Server-side timeout for an individual gRPC request.
     ///
     /// If the handler takes longer than this duration, the server cancels the call.
@@ -155,12 +150,7 @@ impl BlockProducer {
         // any complete or fail, we can shutdown the rest (somewhat) gracefully.
         let mut tasks = tokio::task::JoinSet::new();
 
-        // Launch the gRPC server and wait at the checkpoint for any other components to be in sync.
-        //
-        // This is used to ensure the ntx-builder can subscribe to the mempool events without
-        // playing catch up caused by block-production.
-        //
-        // This is a temporary work-around until the ntx-builder can resync on the fly.
+        // Launch the gRPC server.
         let rpc_id = tasks
             .spawn({
                 let mempool = mempool.clone();
@@ -171,7 +161,6 @@ impl BlockProducer {
                 }
             })
             .id();
-        self.production_checkpoint.wait().await;
 
         let batch_builder_id = tasks
             .spawn({